Solution to Create a Program to Store and Display Student Records

+++ Create a Program to Store and Display Student Records@@67877d61dc107486ad5c2742/code.c
@@ -2,7 +2,7 @@
 struct students{
     int rollnumber;
     int name[50];
-    int marks;
+    float marks;
 };
 int main(){
     int n;
